First Commit
First commit after fixing issue. Issue fix with lf to crlf was the gitignore file was not placed within the project files. Simply doing that fixed it.

Within this commit I turned on the box colliders for all the walls and checked the unchecked boxes on each pickup, only four of them were unchecked